Rewrite BijectiveAutoencoder, moons example
