+++ examples/default_config/moons.yaml
@@ -1,3 +1,5 @@
+epochs: 10
+
 flow:
     transformer: torchnf.transformers.AffineTransform
     net:
@@ -12,13 +14,7 @@ optimizer:
         lr: 0.001
     scheduler: CosineAnnealingLR
     scheduler_init:
-        T_max: 10000
-    submodule: flow
-
-trainer:
-    enable_checkpointing: False
-    max_epochs: 10
-    val_check_interval: 0.25
+        T_max: 10 # should be same as epochs
 
 moons:
     total_size: 6400
