+++ src/torchnf/models/__init__.py
@@ -1,0 +1,1 @@
+from .bijective_autoencoder import BijectiveAutoencoder  # noqa: F401
